+++ tests/conftest.py
@@ -3,9 +3,12 @@ import logging
 import os
 import sys
 
+import pandas as pd
 import pytest
 
 from tradingstrategy.client import Client
+from tradingstrategy.exchange import ExchangeUniverse
+from tradingstrategy.pair import PandasPairUniverse
 
 
 @pytest.fixture(scope="session")
@@ -73,3 +76,35 @@ def logger(request) -> logging.Logger:
     logging.getLogger("asyncio").setLevel(logging.WARNING)
 
     return logger
+
+
+@pytest.fixture(scope='session')
+def default_exchange_universe(persistent_test_client: Client) -> ExchangeUniverse:
+    """Load and construct exchange universe from the live server"""
+    client = persistent_test_client
+    exchange_universe = client.fetch_exchange_universe()
+    return exchange_universe
+
+
+@pytest.fixture(scope='session')
+def default_pairs_df(persistent_test_client: Client) -> pd.DataFrame:
+    """Load pairs data from the live server"""
+    client = persistent_test_client
+    raw_pairs = client.fetch_pair_universe().to_pandas()
+    return raw_pairs
+
+
+@pytest.fixture(scope='session')
+def default_pair_universe(
+    persistent_test_client: Client,
+    default_exchange_universe: ExchangeUniverse,
+    default_pairs_df,
+) -> PandasPairUniverse:
+    """Load and construct pair universe from the live server.
+
+    - Include and index all pairs
+    """
+    raw_pairs = default_pairs_df
+    pair_universe = PandasPairUniverse(raw_pairs, build_index=True, exchange_universe=default_exchange_universe)
+    return pair_universe
+