+++ tradingstrategy/client.py
@@ -53,7 +53,7 @@ from tradingstrategy.environment.config import Configuration
 
 from tradingstrategy.exchange import ExchangeUniverse
 from tradingstrategy.timebucket import TimeBucket
-from tradingstrategy.transport.cache import CachedHTTPTransport, DataNotAvailable
+from tradingstrategy.transport.cache import CachedHTTPTransport, DataNotAvailable, OHLCVCandleType
 
 logger = logging.getLogger(__name__)
 
@@ -275,6 +275,7 @@ class Client(BaseClient):
         start_time: Optional[AnyTimestamp] = None,
         end_time: Optional[AnyTimestamp] = None,
         progress_bar_description: Optional[str] = None,
+        query_type: OHLCVCandleType = OHLCVCandleType.tvl_v1,
     ) -> pd.DataFrame:
         """Fetch TVL/liquidity candles for particular trading pairs.
 
@@ -282,6 +283,8 @@ class Client(BaseClient):
         or few trading pairs. If the number
         of trading pair is small, this download is much more lightweight
         than Parquet dataset download.
+
+        TODO: Upgrade default `query_type` from `tvl_v1` to `tvl_v2`. Lots of backwards breaking changes.
 
         The returned TVL/liquidity data is converted to US dollars by the server.
 
@@ -367,6 +370,7 @@ class Client(BaseClient):
             start_time,
             end_time,
             progress_bar_description=progress_bar_description,
+            query_type=query_type,
         )
 
     def fetch_clmm_liquidity_provision_candles_by_pair_ids(self,