+++ tradingstrategy/transport/cache.py
@@ -44,9 +44,23 @@ class OHLCVCandleType(enum.Enum):
     """Candle types for /candles endpoint
 
     See /candles as https://tradingstrategy.ai/api/explorer/#/Trading%20pair/web_candles
-    ."""
+    """
+
+    #: OHLCV price data
     price = "price"
-    tvl = "tvl"
+
+    #: Contains one-sided liquidity for XYLiquidity Uniswap v2 pairs and some V3 dollar nominated pairs
+    #:
+    #: This is the legacy method.
+    #:
+    tvl_v1 = "tvl"
+
+    #: Contains one-sided quote-token measured, dollar-nominated, TVL for Uniswap v2 and v3 pairs
+    #:
+    #: This is the recommended method.
+    #:
+    tvl_v2 = "tvl2"
+
 
 
 class APIError(Exception):
@@ -695,6 +709,7 @@ class CachedHTTPTransport:
         time_bucket: TimeBucket,
         start_time: Optional[datetime.datetime] = None,
         end_time: Optional[datetime.datetime] = None,
+        query_type: OHLCVCandleType = OHLCVCandleType.tvl_v1,
     ) -> pd.DataFrame:
         """Internal hack to load TVL data for a single pair.
 
@@ -710,14 +725,18 @@ class CachedHTTPTransport:
 
         """
 
+        assert query_type in (OHLCVCandleType.tvl_v1, OHLCVCandleType.tvl_v2,), f"Got: {query_type}"
+
         pair_ids = [pair_id]
+
+        candle_type = query_type.value
 
         cache_fname = self._generate_cache_name(
             pair_ids,
             time_bucket,
             start_time,
             end_time,
-            candle_type="tvl",
+            candle_type=candle_type,
         )
 
         full_fname = self.get_cached_file_path(cache_fname)
@@ -739,13 +758,13 @@ class CachedHTTPTransport:
                     "pair_id": pair_ids[0],
                 }
 
+                params["candle_type"] = query_type.value
+
                 if start_time:
                     params["start"] = start_time.isoformat()
 
                 if end_time:
                     params["end"] = end_time.isoformat()
-
-                params["candle_type"] = OHLCVCandleType.tvl.value
 
                 # Use /candles endpoint to load TVL data
                 pair_candle_map = self.get_json_response(
@@ -794,6 +813,7 @@ class CachedHTTPTransport:
         start_time: Optional[datetime.datetime] = None,
         end_time: Optional[datetime.datetime] = None,
         progress_bar_description: Optional[str] = None,
+        query_type: OHLCVCandleType = OHLCVCandleType.tvl_v1,
     ) -> pd.DataFrame:
         """Load particular set of the TVL candles and cache the result.
 
@@ -850,6 +870,7 @@ class CachedHTTPTransport:
                 time_bucket,
                 start_time,
                 end_time,
+                query_type=query_type,
             )
 
             assert "timestamp" in df.columns, f"Columns lack timestamp: {df.columns}"